backend/requirements.txt
Django>=5.0,<6.0
strawberry-graphql[django]>=0.243.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.4.0
celery>=5.4.0
redis>=5.1.0
django-celery-beat>=2.7.0

# Produção
gunicorn>=22.0.0
whitenoise>=6.7.0
dj-database-url>=2.2.0
